+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change log
 
+## [Version 1.0.4](https://github.com/efremidze/Magnetic/releases/tag/1.0.4)
+Released on 2017-04-06
+
+- Added animation options
+
+## [Version 1.0.3](https://github.com/efremidze/Magnetic/releases/tag/1.0.3)
+Released on 2017-04-01
+
+- Added delegate methods for selection
+- Added documentation
+- Updated Example
+
 ## [Version 1.0.2](https://github.com/efremidze/Magnetic/releases/tag/1.0.2)
 Released on 2017-03-27